refactor(frontend): group free-floating components into theme folders
Move the loose files in components/ into cohesive subfolders
(thesis/, assets/, chart/, notes/, layout/, tags/, search/, settings/,
indicators/, context-menu/, pseudo-etf/), matching the existing
chart//group-table/ convention. Genuinely cross-cutting primitives
(icon-picker, market-status-dot, scanner-view, error-boundary,
period-selector) stay top-level rather than forced into a shared/
junk bucket.

All pseudo-ETF UI is consolidated under pseudo-etf/: the picker plus
the charts pulled out of chart/pseudo-etf/, so the whole feature lives
in one place.

Pure structural move: git mv + @/components import-path rewrites, plus
fixing price-chart's now-sibling relative imports. No logic changes.
Verified with eslint + tsc -b && vite build in the frontend container.
