+++ frontend/src/components/assets/asset-card.tsx
@@ -4,17 +4,17 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Badge } from "@/components/ui/badge"
 import { Skeleton } from "@/components/ui/skeleton"
 import { ContextMenu, ContextMenuTrigger } from "@/components/ui/context-menu"
-import { AssetContextMenuContent } from "@/components/asset-context-menu"
-import { EditAssetDialog } from "@/components/edit-asset-dialog"
-import { NewThesisDialog } from "@/components/new-thesis-dialog"
+import { AssetContextMenuContent } from "@/components/assets/asset-context-menu"
+import { EditAssetDialog } from "@/components/assets/edit-asset-dialog"
+import { NewThesisDialog } from "@/components/thesis/new-thesis-dialog"
 import { MarketStatusDot } from "@/components/market-status-dot"
 import { useAddAssetToThesis } from "@/lib/queries"
-import { DeferredSparkline } from "@/components/sparkline"
-import { TagBadge } from "@/components/tag-badge"
+import { DeferredSparkline } from "@/components/chart/sparkline"
+import { TagBadge } from "@/components/tags/tag-badge"
 import type { AssetType, Quote, TagBrief, SparklinePoint, IndicatorSummary } from "@/lib/api"
 import { formatAssetPrice, formatAssetCompactPrice, changeColor, formatChangePct } from "@/lib/format"
 import { getCardDescriptors, isVisibleAt, type IndicatorDescriptor, type Placement } from "@/lib/indicator-registry"
-import { IndicatorValue } from "@/components/indicator-value"
+import { IndicatorValue } from "@/components/indicators/indicator-value"
 import { usePriceFlash } from "@/lib/use-price-flash"
 import { useSettings } from "@/lib/settings"
