+++ frontend/src/components/assets/earnings-countdown.tsx
