+++ frontend/src/components/assets/stats-panel.tsx
