+++ frontend/src/components/assets/window-selector.tsx
