+++ frontend/src/components/chart/chart-skeleton.tsx
