+++ frontend/src/components/chart/expanded-asset-chart.tsx
