+++ frontend/src/components/chart/intraday-chart.tsx
