+++ frontend/src/components/chart/mini-sparkline.tsx
