+++ frontend/src/components/chart/sparkline.tsx
