+++ frontend/src/components/group-table/table-row.tsx
@@ -4,12 +4,12 @@ import { ChevronRight, ChevronDown } from "lucide-react"
 import { Skeleton } from "@/components/ui/skeleton"
 import { Badge } from "@/components/ui/badge"
 import { ContextMenu, ContextMenuTrigger } from "@/components/ui/context-menu"
-import { EditAssetDialog } from "@/components/edit-asset-dialog"
-import { NewThesisDialog } from "@/components/new-thesis-dialog"
+import { EditAssetDialog } from "@/components/assets/edit-asset-dialog"
+import { NewThesisDialog } from "@/components/thesis/new-thesis-dialog"
 import { useAddAssetToThesis } from "@/lib/queries"
-import { TagBadge } from "@/components/tag-badge"
+import { TagBadge } from "@/components/tags/tag-badge"
 import { MarketStatusDot } from "@/components/market-status-dot"
-import { ExpandedAssetChart } from "@/components/expanded-asset-chart"
+import { ExpandedAssetChart } from "@/components/chart/expanded-asset-chart"
 import type { Asset, Quote, IndicatorSummary } from "@/lib/api"
 import { formatAssetPrice, formatAssetCompactPrice, formatCompactNumber, changeColor, formatChangePct, readableTextColor } from "@/lib/format"
 import {
