+++ frontend/src/components/indicators/indicator-cell.tsx
