+++ frontend/src/components/indicators/indicator-value.tsx
