+++ frontend/src/components/layout/edit-group-dialog.tsx
