+++ frontend/src/components/notes/annotations-list.tsx
