+++ frontend/src/components/notes/note-editor.tsx
