+++ frontend/src/components/pseudo-etf/daily-contribution-chart.tsx
