+++ frontend/src/components/pseudo-etf/index.ts
