+++ frontend/src/components/pseudo-etf/performance-overlay-chart.tsx
