+++ frontend/src/components/pseudo-etf/symbol-legend.tsx
