+++ frontend/src/components/search/search-result-item.tsx
