+++ frontend/src/components/settings/indicator-visibility-editor.tsx
