+++ frontend/src/components/settings/symbol-sources-settings.tsx
