+++ frontend/src/components/settings/visibility-toggle.tsx
