+++ frontend/src/components/thesis/edit-thesis-dialog.tsx
