+++ frontend/src/components/thesis/new-thesis-dialog.tsx
