+++ frontend/src/components/thesis/thesis-member-context-menu.tsx
@@ -1,6 +1,6 @@
 import { FolderPlus, Pencil, Trash2 } from "lucide-react"
-import { ContextActionsContent, type ContextAction } from "@/components/context-actionable"
-import { useThesisMembershipAction, groupTargetActions } from "@/components/menu-actions"
+import { ContextActionsContent, type ContextAction } from "@/components/context-menu/context-actionable"
+import { useThesisMembershipAction, groupTargetActions } from "@/components/context-menu/menu-actions"
 import { useGroups, useAddAssetsToGroup, useRemoveAssetFromThesis } from "@/lib/queries"
 
 interface ThesisMemberContextMenuContentProps {
