+++ frontend/src/pages/asset-detail/chart-section.tsx
@@ -1,6 +1,6 @@
-import { PriceChart } from "@/components/price-chart"
-import { ChartSkeleton } from "@/components/chart-skeleton"
-import { IntradayChart } from "@/components/intraday-chart"
+import { PriceChart } from "@/components/chart/price-chart"
+import { ChartSkeleton } from "@/components/chart/chart-skeleton"
+import { IntradayChart } from "@/components/chart/intraday-chart"
 import { useAssetWindow, useAnnotations } from "@/lib/queries"
 import { useIntraday, useQuote } from "@/lib/quote-stream"
 import type { AssetWindow } from "@/lib/asset-window"
