+++ frontend/src/pages/group-page.tsx
@@ -12,9 +12,9 @@ import { Input } from "@/components/ui/input"
 import { SegmentedControl } from "@/components/ui/segmented-control"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Skeleton } from "@/components/ui/skeleton"
-import { AddSymbolDialog } from "@/components/add-symbol-dialog"
-import { AssetCard } from "@/components/asset-card"
-import { TagFilterPopover } from "@/components/tag-filter-popover"
+import { AddSymbolDialog } from "@/components/assets/add-symbol-dialog"
+import { AssetCard } from "@/components/assets/asset-card"
+import { TagFilterPopover } from "@/components/tags/tag-filter-popover"
 import { useGroup, useGroups, useGroupSparklines, useGroupIndicators, useRemoveAssetFromGroup, useUpdateGroup, useTags, useTheses, usePrefetchAssetDetail, usePrefetchOtherGroups } from "@/lib/queries"
 import { useQuotes } from "@/lib/quote-stream"
 import { buildSortOptions, getScannableDescriptors } from "@/lib/indicator-registry"
@@ -22,11 +22,11 @@ import { useSettings, type AssetTypeFilter, type GroupSortBy, type GroupViewMode
 import { useFilteredSortedAssets, getSortValue, compareSortValues, type SortValue } from "@/lib/use-group-filter"
 import type { Asset, Thesis } from "@/lib/api"
 import { GroupTable } from "@/components/group-table"
-import { AssetContextMenuContent } from "@/components/asset-context-menu"
-import { ThesisGroupedTable } from "@/components/thesis-grouped-table"
+import { AssetContextMenuContent } from "@/components/assets/asset-context-menu"
+import { ThesisGroupedTable } from "@/components/thesis/thesis-grouped-table"
 import { CrosshairTimeSyncProvider } from "@/components/chart/crosshair-time-sync"
 import { ScannerView } from "@/components/scanner-view"
-import { LiveDayView } from "@/components/live-day-view"
+import { LiveDayView } from "@/components/chart/live-day-view"
 
 const SCANNABLE_DESCRIPTORS = getScannableDescriptors()
