+++ frontend/src/pages/pseudo-etf-detail.tsx
@@ -3,15 +3,15 @@ import { useParams, Link, Navigate } from "react-router-dom"
 import { ArrowLeft, UserPlus, Loader2 } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { ConnectedNote } from "@/components/connected-note"
-import { ConnectedAnnotations } from "@/components/connected-annotations"
+import { ConnectedNote } from "@/components/notes/connected-note"
+import { ConnectedAnnotations } from "@/components/notes/connected-annotations"
 import { HoldingsGrid, type HoldingsGridRow } from "@/components/holdings-grid"
-import { AddConstituentPicker } from "@/components/add-constituent-picker"
+import { AddConstituentPicker } from "@/components/pseudo-etf/add-constituent-picker"
 import { CrosshairTimeSyncProvider } from "@/components/chart/crosshair-time-sync"
 import {
   PerformanceOverlayChart,
   DailyContributionChart,
-} from "@/components/chart/pseudo-etf"
+} from "@/components/pseudo-etf"
 import { STACK_COLORS } from "@/lib/chart-utils"
 import { formatChangePct } from "@/lib/format"
 import { useSettings } from "@/lib/settings"
